route: move platform-agnostic route code into route.go
This is shared across Darwin and Linux builds.

Signed-off by: Ian Vernon <ian@cilium.io>
